requirements.txt
Flask==2.3.2
icetk==0.0.7
pandas==2.0.3
python_snippets==0.0.11
requests==2.31.0
streamlit==1.24.1
tqdm==4.65.0
